Refactor hfp components (no more hoc), add list for area journeys.
