+++ src/languages/en.json
@@ -21,6 +21,7 @@
   "sidepanel.tabs.journeys": "Journeys",
   "sidepanel.tabs.timetables": "Timetables",
   "sidepanel.tabs.vehicle_journeys": "Vehicle Journeys",
+  "sidepanel.tabs.area_events": "Area journeys",
 
   "map.stops.arrive": "Arrival",
   "map.stops.depart": "Departure",
