+++ src/languages/fi.json
@@ -21,6 +21,7 @@
   "sidepanel.tabs.journeys": "Lähdöt",
   "sidepanel.tabs.timetables": "Aikataulut",
   "sidepanel.tabs.vehicle_journeys": "Lähtöketju",
+  "sidepanel.tabs.area_events": "Alueen matkat",
 
   "map.stops.arrive": "Saapumisaika",
   "map.stops.depart": "Lähtöaika",
