+++ src/languages/se.json
@@ -21,6 +21,7 @@
   "sidepanel.tabs.journeys": "Resor",
   "sidepanel.tabs.timetables": "Tidtabeller",
   "sidepanel.tabs.vehicle_journeys": "Fordonsresor",
+  "sidepanel.tabs.area_events": "Områdets resor",
 
   "map.stops.arrive": "Anländningstid",
   "map.stops.depart": "Avgångstid",
